refactor(elastic): restructure module based on calculation methods
This commit refactors the 'elastic' module to align its structure with the three canonical methods for calculating elastic constants as described in the literature (e.g., Clavier et al., 2017).

The new structure is organized by methodology (Deformation Method, Fluctuation Method) rather than by temperature, which improves clarity, modularity, and future extensibility.

Changes include:
- Created 'deformation_method' and 'fluctuation_method' sub-packages.
- Moved and renamed 'zeroelasticity.py' and 'finite_temp_elasticity.py' into the 'deformation_method' sub-package.
- Updated '__init__.py' files at all levels to correctly manage the package's public API and internal imports.
- Added placeholder files for future implementation of the fluctuation methods.
- Updated docstrings in the moved files to reflect their new names and paths.
