+++ src/thermoelasticsim/elastic/deformation_method/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+显式形变法子包
+
+.. moduleauthor:: Gilbert Young
+.. created:: 2025-07-08
+.. modified:: 2025-07-08
+.. version:: 4.0.0
+"""
+from .zero_temp import ElasticConstantsWorkflow
+from .finite_temp import FiniteTempElasticityWorkflow
+
+__all__ = ["ElasticConstantsWorkflow", "FiniteTempElasticityWorkflow"]
