+++ src/thermoelasticsim/elastic/fluctuation_method/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+涨落法子包
+
+.. moduleauthor:: Gilbert Young
+.. created:: 2025-07-08
+.. modified:: 2025-07-08
+.. version:: 4.0.0
+"""
+
+# --- 模块内容将在未来实现 ---
+__all__ = []
