+++ src/thermoelasticsim/elastic/fluctuation_method/stress_fluctuation.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+ThermoElasticSim - 应力涨落法模块
+
+.. moduleauthor:: Gilbert Young
+.. created:: 2025-07-08
+.. modified:: 2025-07-08
+.. version:: 4.0.0
+
+该模块将实现基于应力涨落理论的弹性常数计算方法。
+
+此方法通过在NVT系综下长时间模拟，分析应力张量的自发涨落，
+并结合势能对应变的二阶导数（玻恩项），来计算材料的弹性常数。
+
+.. math::
+    C_{ijkl} = \langle C^B_{ijkl} \rangle - \frac{V}{k_B T} \left[ \langle \sigma_{ij} \sigma_{kl} \rangle - \langle \sigma_{ij} \rangle \langle \sigma_{kl} \rangle \right]
+
+参考文献:
+    Clavier, G., et al. (2017). Computation of elastic constants of solids
+    using molecular simulation. Molecular Simulation, 43(13-16), 1133-1145.
+
+.. note::
+    根据 Clavier et al. 的研究，这是最高效且最稳健的有限温计算方法，
+    在MD和MC中均表现良好，且不受控温器参数影响。
+    其主要挑战在于需要计算玻恩项 :math:`C^B`。
+"""
+
+# --- 模块内容将在未来实现 ---
